DailyReport.phpをapp/Modelsに配置
